requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
